refactor(dbHelpers.ts): move listHelpers to dbHelpers and update imports in affected files
